Took missing stuff from multibrand branch

+++ tms/Makefile.py
@@ -338,17 +338,23 @@ def terminal(*a):
 
 @target
 def batch(*a):
+    '''batch|mbatch'''
     assert len(parameters) > 0, 'Which module to run?'
-    batch_module = parameters[0] 
+    batch_module = parameters[0]
+    
     module_base = os.path.basename(batch_module)
+    if 'tenant' in globals():
+        module_base = '{}_{}'.format(module_base,tenant)
+
     cdr_dict = {}
 
-    if os.path.exists('../var/run/%s.pid' % module_base):
-        print('Lockfile %s.pid exists - aborting!' % module_base)
-        sys.exit(5)
-    fd = open('../var/run/%s.pid' % module_base, 'w')
-    fd.write(str(os.getpid()))
-    fd.close()
+    if a[0] == 'batch':
+       if os.path.exists('../var/run/%s.pid' % module_base):
+           print('Lockfile %s.pid exists - aborting!' % module_base)
+           sys.exit(5)
+       fd = open('../var/run/%s.pid' % module_base, 'w')
+       fd.write(str(os.getpid()))
+       fd.close()
 
     args = ['-T', '../var/tmp', '-b', '-p', batch_module + '.p']
 
@@ -384,7 +390,8 @@ def batch(*a):
             cmd.wait()
         except KeyboardInterrupt:
             cmd.send_signal(2)
-    os.unlink('../var/run/%s.pid' % module_base)
+    if a[0] == 'batch':
+      os.unlink('../var/run/%s.pid' % module_base)
     sys.exit(cmd.returncode)
 
 @target
@@ -392,6 +399,9 @@ def idbatch(*a):
     assert len(parameters) > 0, 'Which module to run?'
     batch_module = parameters[0]
     module_base = os.path.basename(batch_module)
+    if 'tenant' in globals():
+        module_base = '{}_{}'.format(module_base,tenant)
+
     cdr_dict = {}
 
     try:
